Move game state classes into states folder
